Attach sessions to all model class instances
Previously, the references were immediately resolved to the module
`manager_api` namespace, but they need to refer to the subclasses which
are made when a business is accessed.

Move forward reference updating into `Business` initialization, and use
that opportunity to tie the references to the subclasses with the
session attached. This required referencing all classes internally via
the module (`model`) so that they would remain forward references until
a business is initialized.