Read object before allowing field update
Since objects automatically read from the API when accessing any
attribute, a user might expect to be able to retrieve a list of objects,
make a modification to each object, and then update them. However, the
objects were empty because they were never fetched. A subsequent call to
`update` would then clear out all attributes of the object in Manager.

This change enforces automatic read whenever an attribute of an object
is set or modified.